+++ src/components/common/StatusRenderer/BoxesRenderer.module.css
