+++ src/components/common/StatusRenderer/index.ts
@@ -1,0 +1,2 @@
+export { default as BoxesRenderer } from './BoxesRenderer';
+export { default as BoxRenderer } from './BoxRenderer';
